# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/admin/order/": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get record of order",
                "description": "Returns a list of order",
                "operationId": "order/list",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/admin/order": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Store order",
                "description": "Store new order.",
                "operationId": "order/storesy",
                "requestBody": {
                    "description": "Array of categories to store",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer_info": {
                                        "properties": {
                                            "customer_id": {
                                                "type": "number",
                                                "example": "2"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "order_details": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "order_id": {
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "medicine_id": {
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "item_price": {
                                                    "type": "number",
                                                    "example": "139"
                                                },
                                                "qty": {
                                                    "type": "number",
                                                    "example": "4"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "order successfully stored"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/admin/order/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": " order single record",
                "description": "order single record",
                "operationId": "order/get{id}",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "pass id of record",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "order detete",
                "description": "delete  order",
                "operationId": "order/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "pass id for order",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Failed operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}